Updated patient table (using generated client)
